feat: Add error handling for Post not found

+++ src/main/java/br/com/felipe/blog/service/PostService.java
@@ -1,6 +1,7 @@
 package br.com.felipe.blog.service;
 
 import br.com.felipe.blog.entity.Post;
+import br.com.felipe.blog.exceptions.ResourceNotFoundException;
 import br.com.felipe.blog.repository.PostRepository;
 import org.springframework.stereotype.Service;
 
@@ -20,7 +21,7 @@ public class PostService {
     }
 
     public Post findById(Long id){
-        return postRepository.findById(id).orElse(null);
+        return postRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Post de id " + id + " não encontrado"));
     }
 
     public List<Post> findAll(){
@@ -28,6 +29,9 @@ public class PostService {
     }
 
     public void deleteById(Long id){
+        if(!postRepository.existsById(id)){
+            throw new ResourceNotFoundException("Post de id " + id + " não encontrado");
+        }
         postRepository.deleteById(id);
     }
 }
